# app/api/campaigns/[campaignId]/intent/interpret/route.ts
import { NextResponse } from "next/server"; import { SLICE_CAMPAIGN_ID,sliceService } from "@/src/server/vertical-slice";
export async function POST(request:Request,{params}:{params:Promise<{campaignId:string}>}){if((await params).campaignId!==SLICE_CAMPAIGN_ID)return NextResponse.json({error:"Campaign not found"},{status:404});const body=await request.json() as {goal?:string;method?:string;stateVersion?:string};if(!body.goal?.trim()||!body.method?.trim())return NextResponse.json({error:"Goal and method are required"},{status:400});const state=sliceService().state(SLICE_CAMPAIGN_ID);if(Number(body.stateVersion??state.stateVersion)!==state.stateVersion)return NextResponse.json({error:"Stage changed; refresh before interpreting"},{status:409});return NextResponse.json({proposalId:crypto.randomUUID(),goal:body.goal.trim(),method:body.method.trim(),difficulty:12,accessibleStakes:["Tide pressure may advance","A mixed outcome may cost vitality","The threshold may remain sealed"],stateVersion:String(state.stateVersion)});}