# contracts/source/c1_contracts/openapi/mythic-ui-c1.yaml
openapi: 3.1.2
info:
  title: Mythic RPG Engine/UI Contract — Checkpoint C1
  version: 0.1.0
  description: Synchronous viewpoint-safe reads and immutable core-loop commands. No operation grants the UI direct canonical mutation authority.
servers:
  - url: https://api.example.invalid/mythic/v1
    description: Placeholder only; no production platform selected
security:
  - bearerAuth: []
paths:
...
  /campaigns/{campaignId}/intent/interpret:
    post:
      operationId: interpretIntent
      parameters:
        - $ref: '#/components/parameters/CampaignId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '../schemas/commands/core-loop.schema.json#/$defs/InterpretIntentCommand' }
      responses:
        '200': { $ref: '#/components/responses/InterpretedProposal' }
        '422': { $ref: '#/components/responses/RejectedReceipt' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    CampaignId: { name: campaignId, in: path, required: true, schema: { type: string } }
...
  responses:
...
    InterpretedProposal:
      description: Versioned, noncanonical interpretation for player review
      content: { application/json: { schema: { $ref: '../schemas/read-models/core-loop.schema.json#/$defs/InterpretedProposal' } } }
...
    RejectedReceipt:
      description: Actor-safe rejection with no hidden validator diagnostics
      content: { application/json: { schema: { $ref: '../schemas/commands/core-loop.schema.json#/$defs/CommandReceipt' } } }
...